Moved helpers to own files; easier to test and find code.
