+++ src/common/tempUmaFunctions.ts
@@ -1,7 +1,6 @@
 // Pulled from @uma/common temporarily, as it has no type declarations.
 // remove any dependencies in here later.
 import EthCrypto from "eth-crypto";
-import { ethers } from "ethers";
 import Web3 from "web3";
 import web3 from "web3";
